arm/altp2m: Extended libxl to activate altp2m on ARM.
The current implementation allows to set the parameter HVM_PARAM_ALTP2M.
This parameter allows further usage of altp2m on ARM. For this, we
define an additional altp2m field for PV domains as part of the
libxl_domain_type struct. This field can be set only on ARM systems
through the altp2mpvh switch in the domain's configuration file (i.e.
set altp2mpvh=1).

This commit addresses and solves the issue mentioned in
27772074c780173bfaad05267eba9b01a329bfac.